chore: isolate code from x/tools
